rust cli: split commands/common.rs into source, parse, render modules
The commands/common.rs grab-bag mixed remote/local input loading, MCAP
parsing into ParsedMcap, and display/value-formatting helpers in one
1,871-line module that was not itself a command. Split it into three
focused crate-root modules:

- source: input loading + remote object-store/HTTP transports
- parse:  ParsedMcap and exact-record parsing
- render: table/byte/time formatting and small CLI value parsers

Also move the shared parse_timestamp_or_nanos helper (used by both
add attachment and filter) out of the add_attachment command and into
render, removing the cross-command import. No logic changes.
