+++ rust/cli/src/commands/add_attachment.rs
@@ -6,6 +6,7 @@ use anyhow::{Context, Result};
 use crate::cli::AddAttachmentCommand;
 use crate::commands::add_common::{self, AttachmentToAdd};
 use crate::context::CommandContext;
+use crate::render::parse_timestamp_or_nanos;
 
 pub fn run(_ctx: &CommandContext, args: AddAttachmentCommand) -> Result<()> {
     let attachment_data = fs::read(&args.attachment_file).with_context(|| {
@@ -54,23 +55,6 @@ pub fn run(_ctx: &CommandContext, args: AddAttachmentCommand) -> Result<()> {
     Ok(())
 }
 
-pub(crate) fn parse_timestamp_or_nanos(value: &str) -> Result<u64> {
-    if let Ok(nanos) = value.parse::<u64>() {
-        return Ok(nanos);
-    }
-
-    let parsed = chrono::DateTime::parse_from_rfc3339(value)
-        .with_context(|| format!("failed to parse timestamp '{value}'"))?;
-    let seconds = parsed.timestamp();
-    anyhow::ensure!(seconds >= 0, "timestamp is before unix epoch: '{value}'");
-    let seconds = seconds as u64;
-    let nanos = parsed.timestamp_subsec_nanos() as u64;
-    seconds
-        .checked_mul(1_000_000_000)
-        .and_then(|v| v.checked_add(nanos))
-        .with_context(|| format!("timestamp is out of range: '{value}'"))
-}
-
 fn system_time_to_nanos(time: SystemTime) -> Result<u64> {
     let duration = time
         .duration_since(UNIX_EPOCH)
@@ -80,22 +64,4 @@ fn system_time_to_nanos(time: SystemTime) -> Result<u64> {
         .checked_mul(1_000_000_000)
         .and_then(|v| v.checked_add(duration.subsec_nanos() as u64))
         .context("timestamp is out of range")
-}
-
-#[cfg(test)]
-mod tests {
-    use super::parse_timestamp_or_nanos;
-
-    #[test]
-    fn parses_nanos_or_rfc3339() {
-        assert_eq!(parse_timestamp_or_nanos("123").expect("nanos"), 123);
-        let ts = parse_timestamp_or_nanos("2023-07-25T15:27:30.132545471Z").expect("rfc3339");
-        assert_eq!(ts, 1_690_298_850_132_545_471);
-    }
-
-    #[test]
-    fn rejects_invalid_timestamp() {
-        let err = parse_timestamp_or_nanos("not-a-time").expect_err("invalid time should fail");
-        assert!(err.to_string().contains("failed to parse timestamp"));
-    }
 }
