+++ rust/cli/src/commands/compress.rs
@@ -15,6 +15,6 @@ pub fn run(ctx: &CommandContext, args: CompressCommand) -> Result<()> {
         .include_attachments(true);
     filter::run_transcode(
         options,
-        crate::commands::common::SourceOptions::new(ctx.allow_remote_scan()),
+        crate::source::SourceOptions::new(ctx.allow_remote_scan()),
     )
 }
