+++ rust/cli/src/commands/list_channels.rs
@@ -1,15 +1,15 @@
 use anyhow::Result;
 
 use crate::cli::ListChannelsCommand;
-use crate::commands::common;
 use crate::context::CommandContext;
+use crate::{render, source};
 
 pub fn run(ctx: &CommandContext, args: ListChannelsCommand) -> Result<()> {
-    let parsed = common::parse_mcap_from_path(
+    let parsed = source::parse_mcap_from_path(
         &args.file,
-        common::SourceOptions::new(ctx.allow_remote_scan()),
+        source::SourceOptions::new(ctx.allow_remote_scan()),
     )?;
-    common::print_table(&render_channel_rows(&parsed.channels)?);
+    render::print_table(&render_channel_rows(&parsed.channels)?);
     Ok(())
 }
