+++ rust/cli/src/source.rs
@@ -1,6 +1,3 @@
-use std::borrow::Cow;
-use std::collections::HashMap;
-use std::fmt::Write as _;
 use std::io::{IsTerminal as _, Read as _, SeekFrom};
 use std::path::Path;
 use std::sync::Arc;
@@ -16,6 +13,9 @@ use object_store::{
 };
 use tempfile::NamedTempFile;
 use url::Url;
+
+use crate::parse::{self, ParsedMcap};
+use crate::render::human_bytes;
 
 pub const PLEASE_REDIRECT: &str =
     "Binary output can screw up your terminal. Supply -o or redirect to a file or pipe";
@@ -131,23 +131,6 @@ impl std::io::Read for StreamingInput {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ParsedSchema {
-    pub header: records::SchemaHeader,
-    pub data: Vec<u8>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct ParsedMcap {
-    pub header: Option<records::Header>,
-    pub statistics: Option<records::Statistics>,
-    pub channels: std::collections::BTreeMap<u16, records::Channel>,
-    pub schemas: std::collections::BTreeMap<u16, ParsedSchema>,
-    pub chunk_indexes: Vec<records::ChunkIndex>,
-    pub attachment_indexes: Vec<records::AttachmentIndex>,
-    pub metadata_indexes: Vec<records::MetadataIndex>,
-}
-
 pub fn map_file(path: &Path) -> anyhow::Result<Mmap> {
     let file =
         std::fs::File::open(path).with_context(|| format!("couldn't open '{}'", path.display()))?;
@@ -171,7 +154,7 @@ pub fn parse_mcap_from_path(path: &Path, options: SourceOptions) -> Result<Parse
                     })?
                 {
                     let header = read_header_from_seekable(&mut reader)?;
-                    return Ok(parsed_mcap_from_summary_ref(header, &summary));
+                    return Ok(parse::parsed_mcap_from_summary_ref(header, &summary));
                 }
                 if !options.allow_remote_scan {
                     bail!(
@@ -194,12 +177,12 @@ pub fn parse_mcap_from_path(path: &Path, options: SourceOptions) -> Result<Parse
         let mut source = open_seekable_mcap_source(path)?;
         let header = read_header_from_seekable(&mut source)?;
         if let Some(summary) = read_summary_from_seekable(&mut source)? {
-            return Ok(parsed_mcap_from_summary_ref(header, &summary));
+            return Ok(parse::parsed_mcap_from_summary_ref(header, &summary));
         }
     }
 
     let mcap = load_path(path, options)?;
-    parse_mcap(&mcap)
+    parse::parse_mcap(&mcap)
 }
 
 pub fn load_path(path: &Path, options: SourceOptions) -> Result<InputData> {
@@ -771,121 +754,7 @@ fn read_summary_from_remote(
     if summary_bytes.len() != summary_len {
         return Err(mcap::McapError::UnexpectedEof.into());
     }
-    Ok(Some(parse_summary_section(&summary_bytes)?))
-}
-
-// TODO: keep this in sync with mcap::sans_io::SummaryReader and mcap::read::ChannelAccumulator.
-// A future mcap crate range-summary API should replace this CLI-local parser.
-fn parse_summary_section(summary: &[u8]) -> Result<mcap::Summary> {
-    let mut out = mcap::Summary::default();
-    let mut schemas = HashMap::<u16, Arc<mcap::Schema<'static>>>::new();
-
-    for record in mcap::read::LinearReader::sans_magic(summary) {
-        match record? {
-            Record::AttachmentIndex(index) => out.attachment_indexes.push(index),
-            Record::MetadataIndex(index) => out.metadata_indexes.push(index),
-            Record::Statistics(statistics) => out.stats = Some(statistics),
-            Record::ChunkIndex(index) => out.chunk_indexes.push(index),
-            Record::Schema { header, data } => {
-                if header.id == 0 {
-                    return Err(mcap::McapError::InvalidSchemaId.into());
-                }
-                let schema = Arc::new(mcap::Schema {
-                    id: header.id,
-                    name: header.name,
-                    encoding: header.encoding,
-                    data: Cow::Owned(data.into_owned()),
-                });
-                match schemas.entry(schema.id) {
-                    std::collections::hash_map::Entry::Occupied(entry) => {
-                        let existing = entry.get();
-                        if existing.name != schema.name
-                            || existing.encoding != schema.encoding
-                            || existing.data.as_ref() != schema.data.as_ref()
-                        {
-                            return Err(
-                                mcap::McapError::ConflictingSchemas(schema.name.clone()).into()
-                            );
-                        }
-                    }
-                    std::collections::hash_map::Entry::Vacant(entry) => {
-                        entry.insert(schema);
-                    }
-                }
-            }
-            Record::Channel(channel) => {
-                let schema = if channel.schema_id == 0 {
-                    None
-                } else {
-                    Some(schemas.get(&channel.schema_id).cloned().ok_or_else(|| {
-                        mcap::McapError::UnknownSchema(channel.topic.clone(), channel.schema_id)
-                    })?)
-                };
-                let resolved = Arc::new(mcap::Channel {
-                    id: channel.id,
-                    topic: channel.topic,
-                    schema,
-                    message_encoding: channel.message_encoding,
-                    metadata: channel.metadata,
-                });
-                match out.channels.entry(resolved.id) {
-                    std::collections::hash_map::Entry::Occupied(entry) => {
-                        let existing = entry.get();
-                        if existing.topic != resolved.topic
-                            || existing.schema.as_ref().map(|schema| schema.id)
-                                != resolved.schema.as_ref().map(|schema| schema.id)
-                            || existing.message_encoding != resolved.message_encoding
-                            || existing.metadata != resolved.metadata
-                        {
-                            return Err(mcap::McapError::ConflictingChannels(
-                                resolved.topic.clone(),
-                            )
-                            .into());
-                        }
-                    }
-                    std::collections::hash_map::Entry::Vacant(entry) => {
-                        entry.insert(resolved);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    out.schemas = schemas;
-    Ok(out)
-}
-
-// TODO: keep these exact-record parsers in sync with mcap::read::metadata and
-// mcap::read::attachment. They duplicate the mcap crate helpers so remote range callers can
-// parse owned records without holding a full-file byte slice alive.
-pub(crate) fn parse_metadata_record(bytes: &[u8]) -> Result<mcap::records::Metadata> {
-    let mut reader = mcap::read::LinearReader::sans_magic(bytes);
-    let metadata = match reader.next().ok_or(mcap::McapError::BadIndex)?? {
-        mcap::records::Record::Metadata(metadata) => metadata,
-        _ => return Err(mcap::McapError::BadIndex.into()),
-    };
-    if reader.next().is_some() {
-        return Err(mcap::McapError::BadIndex.into());
-    }
-    Ok(metadata)
-}
-
-pub(crate) fn parse_attachment_record(bytes: &[u8]) -> Result<mcap::Attachment<'static>> {
-    let mut reader = mcap::read::LinearReader::sans_magic(bytes);
-    let attachment = match reader.next().ok_or(mcap::McapError::BadIndex)?? {
-        mcap::records::Record::Attachment { header, data, .. } => mcap::Attachment {
-            log_time: header.log_time,
-            create_time: header.create_time,
-            name: header.name,
-            media_type: header.media_type,
-            data: Cow::Owned(data.into_owned()),
-        },
-        _ => return Err(mcap::McapError::BadIndex.into()),
-    };
-    if reader.next().is_some() {
-        return Err(mcap::McapError::BadIndex.into());
-    }
-    Ok(attachment)
+    Ok(Some(parse::parse_summary_section(&summary_bytes)?))
 }
 
 fn read_summary_from_seekable(
@@ -934,250 +803,6 @@ fn read_header_from_seekable(
     Ok(None)
 }
 
-fn parsed_mcap_from_summary_ref(
-    header: Option<records::Header>,
-    summary: &mcap::Summary,
-) -> ParsedMcap {
-    let mut out = ParsedMcap {
-        header,
-        statistics: summary.stats.clone(),
-        channels: std::collections::BTreeMap::new(),
-        schemas: std::collections::BTreeMap::new(),
-        chunk_indexes: summary.chunk_indexes.clone(),
-        attachment_indexes: summary.attachment_indexes.clone(),
-        metadata_indexes: summary.metadata_indexes.clone(),
-    };
-
-    for schema in summary.schemas.values() {
-        let schema = schema.as_ref();
-        out.schemas.insert(
-            schema.id,
-            ParsedSchema {
-                header: records::SchemaHeader {
-                    id: schema.id,
-                    name: schema.name.clone(),
-                    encoding: schema.encoding.clone(),
-                },
-                data: schema.data.clone().into_owned(),
-            },
-        );
-    }
-
-    for channel in summary.channels.values() {
-        let channel = channel.as_ref();
-        out.channels.insert(
-            channel.id,
-            records::Channel {
-                id: channel.id,
-                schema_id: channel.schema.as_ref().map(|schema| schema.id).unwrap_or(0),
-                topic: channel.topic.clone(),
-                message_encoding: channel.message_encoding.clone(),
-                metadata: channel.metadata.clone(),
-            },
-        );
-    }
-
-    out
-}
-
-pub fn parse_mcap(mcap: &[u8]) -> Result<ParsedMcap> {
-    let header = read_header(mcap)?;
-    if let Some(parsed_from_summary) = parse_mcap_from_summary(mcap, header.clone())? {
-        return Ok(parsed_from_summary);
-    }
-
-    eprintln!(
-        "Warning: summary section not available; full scan may be slow. Run `mcap doctor` for details."
-    );
-    parse_mcap_linear(mcap, header)
-}
-
-pub(crate) fn read_header(mcap: &[u8]) -> Result<Option<records::Header>> {
-    let mut reader = mcap::read::LinearReader::new(mcap)?;
-    match reader.next() {
-        Some(Ok(Record::Header(header))) => Ok(Some(header)),
-        Some(Ok(_)) | None => Ok(None),
-        Some(Err(err)) => Err(err.into()),
-    }
-}
-
-fn parse_mcap_from_summary(
-    mcap: &[u8],
-    header: Option<records::Header>,
-) -> Result<Option<ParsedMcap>> {
-    let Some(summary) = mcap::Summary::read(mcap)? else {
-        return Ok(None);
-    };
-    Ok(Some(parsed_mcap_from_summary_ref(header, &summary)))
-}
-
-fn parse_mcap_linear(mcap: &[u8], header: Option<records::Header>) -> Result<ParsedMcap> {
-    let mut out = ParsedMcap {
-        header,
-        ..ParsedMcap::default()
-    };
-    for record in mcap::read::LinearReader::new(mcap)? {
-        let record = record?;
-        if let Record::Chunk { header, data } = record {
-            for nested_record in mcap::read::ChunkReader::new(header, data.as_ref())? {
-                collect_record(&mut out, nested_record?)?;
-            }
-        } else {
-            collect_record(&mut out, record)?;
-        }
-    }
-
-    Ok(out)
-}
-
-fn collect_record(out: &mut ParsedMcap, record: Record<'_>) -> Result<()> {
-    match record {
-        Record::Header(header) => {
-            if let Some(existing) = &out.header {
-                if existing != &header {
-                    bail!("conflicting MCAP header records");
-                }
-            } else {
-                out.header = Some(header);
-            }
-        }
-        Record::Statistics(statistics) => {
-            out.statistics = Some(statistics);
-        }
-        Record::Channel(channel) => {
-            if let Some(existing) = out.channels.get(&channel.id) {
-                if existing != &channel {
-                    bail!("conflicting channel definition for id {}", channel.id);
-                }
-            } else {
-                out.channels.insert(channel.id, channel);
-            }
-        }
-        Record::Schema { header, data } => {
-            let schema = ParsedSchema {
-                header,
-                data: data.into_owned(),
-            };
-            if let Some(existing) = out.schemas.get(&schema.header.id) {
-                if existing != &schema {
-                    bail!("conflicting schema definition for id {}", schema.header.id);
-                }
-            } else {
-                out.schemas.insert(schema.header.id, schema);
-            }
-        }
-        Record::ChunkIndex(index) => out.chunk_indexes.push(index),
-        Record::AttachmentIndex(index) => out.attachment_indexes.push(index),
-        Record::MetadataIndex(index) => out.metadata_indexes.push(index),
-        _ => {}
-    }
-    Ok(())
-}
-
-pub fn decimal_time(t: u64) -> String {
-    format!("{}.{:09}", t / 1_000_000_000, t % 1_000_000_000)
-}
-
-pub fn raw_time(t: u64) -> String {
-    t.to_string()
-}
-
-pub fn write_raw_time(writer: &mut impl std::io::Write, t: u64) -> std::io::Result<()> {
-    write!(writer, "{t}")
-}
-
-pub fn formatted_time(t: u64) -> String {
-    let seconds = (t / 1_000_000_000) as i64;
-    let nanos = (t % 1_000_000_000) as u32;
-    match chrono::DateTime::from_timestamp(seconds, nanos) {
-        Some(dt) => format!("{} ({})", format_rfc3339_trimmed(dt), decimal_time(t)),
-        None => decimal_time(t),
-    }
-}
-
-pub fn human_bytes(num_bytes: u64) -> String {
-    let prefixes = ["B", "KiB", "MiB", "GiB"];
-    for (index, prefix) in prefixes.iter().enumerate() {
-        let displayed = num_bytes as f64 / 1024f64.powi(index as i32);
-        if displayed <= 1024.0 {
-            return format!("{displayed:.2} {prefix}");
-        }
-    }
-
-    let last = prefixes.len() - 1;
-    let displayed = num_bytes as f64 / 1024f64.powi(last as i32);
-    format!("{displayed:.2} {}", prefixes[last])
-}
-
-pub fn parse_output_compression(value: &str) -> Result<Option<mcap::Compression>> {
-    match value {
-        "zstd" => Ok(Some(mcap::Compression::Zstd)),
-        "lz4" => Ok(Some(mcap::Compression::Lz4)),
-        "none" | "" => Ok(None),
-        _ => bail!(
-            "unrecognized compression format '{value}': valid options are 'lz4', 'zstd', or 'none'"
-        ),
-    }
-}
-
-fn format_rfc3339_trimmed(dt: chrono::DateTime<chrono::Utc>) -> String {
-    let rendered = dt.to_rfc3339_opts(chrono::SecondsFormat::Nanos, true);
-    let Some(without_z) = rendered.strip_suffix('Z') else {
-        return rendered;
-    };
-
-    let Some((prefix, fractional)) = without_z.split_once('.') else {
-        return rendered;
-    };
-
-    let trimmed = fractional.trim_end_matches('0');
-    if trimmed.is_empty() {
-        format!("{prefix}Z")
-    } else {
-        format!("{prefix}.{trimmed}Z")
-    }
-}
-
-pub fn format_table(rows: &[Vec<String>]) -> String {
-    if rows.is_empty() {
-        return String::new();
-    }
-
-    let mut widths = vec![0usize; rows[0].len()];
-    for row in rows {
-        for (idx, value) in row.iter().enumerate() {
-            widths[idx] = widths[idx].max(value.chars().count());
-        }
-    }
-
-    let mut out = String::new();
-    let last_col_idx = rows[0].len().saturating_sub(1);
-    for row in rows {
-        let mut line = String::new();
-        for (idx, value) in row.iter().enumerate() {
-            if idx > 0 {
-                line.push('\t');
-            }
-            if idx == last_col_idx {
-                line.push_str(value);
-            } else {
-                let width = widths[idx];
-                let _ = write!(&mut line, "{value:<width$}");
-            }
-        }
-        let _ = writeln!(&mut out, "{line}");
-    }
-    out
-}
-
-pub fn print_table(rows: &[Vec<String>]) {
-    let rendered = format_table(rows);
-    if rendered.is_empty() {
-        return;
-    }
-    print!("{rendered}");
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::BTreeMap;
@@ -1187,10 +812,8 @@ mod tests {
     use std::sync::Arc;
     use std::thread;
 
-    use super::{
-        decimal_time, format_table, formatted_time, human_bytes, load_path, parse_mcap,
-        parse_mcap_from_summary, print_table, write_raw_time,
-    };
+    use super::load_path;
+    use crate::render::human_bytes;
     use mcap::records;
     use object_store::ObjectStoreExt;
 
@@ -1441,9 +1064,9 @@ mod tests {
         )
         .expect_err("oversized metadata range total should require scan opt-in");
         assert!(err.to_string().contains("metadata records"));
-        assert!(err.to_string().contains(&super::human_bytes(
-            super::MAX_REMOTE_METADATA_BYTES_WITHOUT_SCAN
-        )));
+        assert!(err
+            .to_string()
+            .contains(&human_bytes(super::MAX_REMOTE_METADATA_BYTES_WITHOUT_SCAN)));
         assert!(err.to_string().contains("--allow-remote-scan"));
     }
 
@@ -1680,192 +1303,5 @@ mod tests {
             .expect("non-range HTTP input should materialize with scan opt-in");
 
         assert!(parsed.channels.contains_key(&channel_id));
-    }
-
-    #[test]
-    fn table_printer_handles_empty_input() {
-        print_table(&[]);
-        assert!(format_table(&[]).is_empty());
-    }
-
-    #[test]
-    fn table_formatter_aligns_columns() {
-        let rows = vec![
-            vec!["id".to_string(), "topic".to_string()],
-            vec!["7".to_string(), "/foo".to_string()],
-            vec!["12".to_string(), "/barbaz".to_string()],
-        ];
-        let rendered = format_table(&rows);
-        let lines: Vec<&str> = rendered.lines().collect();
-        assert_eq!(lines.len(), 3);
-        assert!(lines[0].starts_with("id"));
-        assert!(lines[1].contains('\t'));
-        assert!(lines[2].contains("/barbaz"));
-    }
-
-    #[test]
-    fn formatted_time_includes_rfc3339_and_decimal() {
-        assert_eq!(
-            formatted_time(1_000_000_000),
-            "1970-01-01T00:00:01Z (1.000000000)"
-        );
-        assert_eq!(decimal_time(1_234_567_890), "1.234567890");
-        assert_eq!(
-            formatted_time(1_234_567_890),
-            "1970-01-01T00:00:01.23456789Z (1.234567890)"
-        );
-    }
-
-    #[test]
-    fn raw_time_is_unformatted_nanoseconds() {
-        assert_eq!(super::raw_time(1_234_567_890), "1234567890");
-    }
-
-    #[test]
-    fn write_raw_time_writes_unformatted_nanoseconds() {
-        let mut out = Vec::new();
-        write_raw_time(&mut out, 1_234_567_890).expect("should write raw time");
-        assert_eq!(
-            String::from_utf8(out).expect("raw time output should be utf8"),
-            "1234567890"
-        );
-    }
-
-    #[test]
-    fn table_formatter_omits_trailing_whitespace() {
-        let rows = vec![
-            vec!["col1".to_string(), "col2".to_string()],
-            vec!["a".to_string(), "b".to_string()],
-        ];
-        let rendered = format_table(&rows);
-        for line in rendered.lines() {
-            assert!(!line.ends_with(' '));
-            assert!(!line.ends_with('\t'));
-        }
-    }
-
-    #[test]
-    fn human_bytes_scales_units() {
-        assert_eq!(human_bytes(2), "2.00 B");
-        assert_eq!(human_bytes(2 * 1024), "2.00 KiB");
-    }
-
-    #[test]
-    fn parse_output_compression_supports_known_values() {
-        assert!(matches!(
-            super::parse_output_compression("zstd").expect("zstd should parse"),
-            Some(mcap::Compression::Zstd)
-        ));
-        assert!(matches!(
-            super::parse_output_compression("lz4").expect("lz4 should parse"),
-            Some(mcap::Compression::Lz4)
-        ));
-        assert!(super::parse_output_compression("none")
-            .expect("none should parse")
-            .is_none());
-        assert!(super::parse_output_compression("")
-            .expect("empty should parse")
-            .is_none());
-    }
-
-    #[test]
-    fn parse_output_compression_rejects_unknown_values() {
-        let err =
-            super::parse_output_compression("snappy").expect_err("unknown compression should fail");
-        assert!(err
-            .to_string()
-            .contains("unrecognized compression format 'snappy'"));
-    }
-
-    #[test]
-    fn parse_mcap_collects_channels_and_schemas() {
-        let mut buffer = Vec::new();
-        let (schema_id, channel_id) = {
-            let mut writer = mcap::Writer::new(std::io::Cursor::new(&mut buffer)).expect("writer");
-            let schema_id = writer
-                .add_schema("demo_schema", "jsonschema", br#"{"type":"object"}"#)
-                .expect("schema");
-            let channel_id = writer
-                .add_channel(schema_id, "/demo", "json", &BTreeMap::new())
-                .expect("channel");
-            writer
-                .write_to_known_channel(
-                    &records::MessageHeader {
-                        channel_id,
-                        sequence: 1,
-                        log_time: 10,
-                        publish_time: 11,
-                    },
-                    br#"{"k":"v"}"#,
-                )
-                .expect("write message");
-            writer.finish().expect("finish writer");
-            (schema_id, channel_id)
-        };
-
-        let parsed = parse_mcap(&buffer).expect("parse mcap");
-        assert!(parsed.header.is_some());
-        assert!(parsed.channels.contains_key(&channel_id));
-        assert!(parsed.schemas.contains_key(&schema_id));
-    }
-
-    #[test]
-    fn parse_mcap_falls_back_for_summaryless_files() {
-        let mut buffer = Vec::new();
-        let (schema_id, channel_id) = {
-            let mut writer = mcap::WriteOptions::new()
-                .emit_summary_records(false)
-                .emit_summary_offsets(false)
-                .create(std::io::Cursor::new(&mut buffer))
-                .expect("writer");
-            let schema_id = writer
-                .add_schema("demo_schema", "jsonschema", br#"{"type":"object"}"#)
-                .expect("schema");
-            let channel_id = writer
-                .add_channel(schema_id, "/demo", "json", &BTreeMap::new())
-                .expect("channel");
-            writer
-                .write_to_known_channel(
-                    &records::MessageHeader {
-                        channel_id,
-                        sequence: 1,
-                        log_time: 10,
-                        publish_time: 11,
-                    },
-                    br#"{"k":"v"}"#,
-                )
-                .expect("write message");
-            writer.finish().expect("finish writer");
-            (schema_id, channel_id)
-        };
-
-        let parsed = parse_mcap(&buffer).expect("parse mcap");
-        assert!(parsed.header.is_some());
-        assert!(parsed.channels.contains_key(&channel_id));
-        assert!(parsed.schemas.contains_key(&schema_id));
-    }
-
-    #[test]
-    fn parse_mcap_from_summary_accepts_empty_summary() {
-        let mut buffer = Vec::new();
-        {
-            let mut writer = mcap::Writer::new(std::io::Cursor::new(&mut buffer)).expect("writer");
-            writer.finish().expect("finish writer");
-        }
-
-        let parsed = parse_mcap_from_summary(&buffer, None).expect("parse from summary");
-        assert!(parsed.is_some());
-        let parsed = parsed.expect("parsed summary output");
-        if let Some(stats) = &parsed.statistics {
-            assert_eq!(stats.message_count, 0);
-            assert_eq!(stats.channel_count, 0);
-            assert_eq!(stats.attachment_count, 0);
-            assert_eq!(stats.metadata_count, 0);
-        }
-        assert!(parsed.channels.is_empty());
-        assert!(parsed.schemas.is_empty());
-        assert!(parsed.chunk_indexes.is_empty());
-        assert!(parsed.attachment_indexes.is_empty());
-        assert!(parsed.metadata_indexes.is_empty());
     }
 }
